+++ CHANGES.md
@@ -3,7 +3,7 @@
 ## 1.4.0 (TBD)
 ### Features
 * `cmd/validator` will now check that command binary's *will be* downloaded per platform, where previously it would only check if they are available under their respective bundle URL.
-* Added new field `IsUpdateMandatory` to [deployment-config](docs/deployment-config.md), which allows to deny the user the launch of the application .
+* Added new field `IsUpdateMandatory` to [deployment-config](docs/deployment-config.md), which allows to deny the user the launch of the application when system bundles require changes.
 ### Fixes
 * Fixed changes to the launcher icon not being applied after a self-update on Linux.
 * `cmd/validator`: fix only the first missing bundle URL being reported.